helpers/time unit-tests consolidate
